Организация кода. Разбиение кода на модули
